refactor pages, add animate directive
