+++ src/directives/animate/animate.scss
@@ -3,6 +3,10 @@
   animation-duration: 600ms;
 }
 
+.pulse {
+  animation-name: pulse;
+}
+
 @keyframes shake {
   from,
   to {
@@ -25,5 +29,19 @@
   80% {
     -webkit-transform: translate3d(10px, 0, 0);
     transform: translate3d(10px, 0, 0);
+  }
+}
+
+@keyframes pulse {
+  from {
+    transform: scale3d(1, 1, 1);
+  }
+
+  50% {
+    transform: scale3d(1.05, 1.05, 1.05);
+  }
+
+  to {
+    transform: scale3d(1, 1, 1);
   }
 }
