+++ src/pages/pin/pin-modal/pin-modal.scss
@@ -60,14 +60,14 @@ page-pin {
         min-height: 340px;
       }
     }
-    .block-code {
-      display: flex;
-      justify-content: space-between;
-      max-width: 145px;
-      margin: auto;
-      width: 50%;
-      padding-top: 33px;
-    }
+    // .block-code {
+    //   display: flex;
+    //   justify-content: space-between;
+    //   max-width: 145px;
+    //   margin: auto;
+    //   width: 50%;
+    //   padding-top: 33px;
+    // }
     .block-buttons {
       width: 100%;
       .row {
@@ -109,19 +109,19 @@ page-pin {
         }
       }
     }
-    .circle {
-      $circle-diameter: 18px;
-      $transition-duration: 150ms;
-      border-radius: 50%;
-      transition: background-color $transition-duration ease,
-        border $transition-duration ease;
-      border: 1px solid color($colors, light-grey);
-      height: $circle-diameter;
-      width: $circle-diameter;
-    }
-    .filled {
-      background-color: $color-secondary;
-      border: 1px solid $color-secondary;
-    }
+    // .circle {
+    //   $circle-diameter: 18px;
+    //   $transition-duration: 150ms;
+    //   border-radius: 50%;
+    //   transition: background-color $transition-duration ease,
+    //     border $transition-duration ease;
+    //   border: 1px solid color($colors, light-grey);
+    //   height: $circle-diameter;
+    //   width: $circle-diameter;
+    // }
+    // .filled {
+    //   background-color: $color-secondary;
+    //   border: 1px solid $color-secondary;
+    // }
   }
 }
