+++ src/pages/settings/lock/lock.ts
@@ -3,7 +3,7 @@ import { TranslateService } from '@ngx-translate/core';
 import { Events } from 'ionic-angular';
 
 // pages
-import { PinModalPage } from '../../pin/pin';
+import { PinModalPage } from '../../pin/pin-modal/pin-modal';
 
 // providers
 import { ConfigProvider } from '../../../providers/config/config';
@@ -14,11 +14,15 @@ import * as _ from 'lodash';
 
 @Component({
   selector: 'page-lock',
-  templateUrl: 'lock.html',
+  templateUrl: 'lock.html'
 })
 export class LockPage {
-
-  public options: Array<{ label: string, method: string, enabled: boolean, disabled: boolean }> = [];
+  public options: Array<{
+    label: string;
+    method: string;
+    enabled: boolean;
+    disabled: boolean;
+  }> = [];
   public lockOptions: any;
   public needsBackupMsg: string;
 
@@ -40,20 +44,44 @@ export class LockPage {
         {
           label: this.translate.instant('Disabled'),
           method: 'disabled',
-          enabled: !this.lockOptions.method || (this.lockOptions.method && this.lockOptions.method.toLowerCase() == 'disabled' ? true : false),
+          enabled:
+            !this.lockOptions.method ||
+            (this.lockOptions.method &&
+            this.lockOptions.method.toLowerCase() == 'disabled'
+              ? true
+              : false),
           disabled: false
         },
         {
           label: this.translate.instant('PIN'),
           method: 'pin',
-          enabled: this.lockOptions.method && this.lockOptions.method.toLowerCase() == 'pin' ? true : false,
-          disabled: needsBackup || (this.lockOptions.method && this.lockOptions.method.toLowerCase() == 'fingerprint' ? true : false)
+          enabled:
+            this.lockOptions.method &&
+            this.lockOptions.method.toLowerCase() == 'pin'
+              ? true
+              : false,
+          disabled:
+            needsBackup ||
+            (this.lockOptions.method &&
+            this.lockOptions.method.toLowerCase() == 'fingerprint'
+              ? true
+              : false)
         },
         {
           label: this.translate.instant('Fingerprint'),
           method: 'fingerprint',
-          enabled: this.lockOptions.method && this.lockOptions.method.toLowerCase() == 'fingerprint' ? true : false,
-          disabled: !isAvailable || needsBackup || (this.lockOptions.method && this.lockOptions.method.toLowerCase() == 'pin' ? true : false)
+          enabled:
+            this.lockOptions.method &&
+            this.lockOptions.method.toLowerCase() == 'fingerprint'
+              ? true
+              : false,
+          disabled:
+            !isAvailable ||
+            needsBackup ||
+            (this.lockOptions.method &&
+            this.lockOptions.method.toLowerCase() == 'pin'
+              ? true
+              : false)
         }
       ];
     });
@@ -65,8 +93,16 @@ export class LockPage {
         this.openPinModal('pinSetUp');
         break;
       case 'disabled':
-        if (this.lockOptions.method && this.lockOptions.method.toLowerCase() == 'pin') this.openPinModal('removeLock');
-        if (this.lockOptions.method && this.lockOptions.method.toLowerCase() == 'fingerprint') this.removeFingerprint();
+        if (
+          this.lockOptions.method &&
+          this.lockOptions.method.toLowerCase() == 'pin'
+        )
+          this.openPinModal('removeLock');
+        if (
+          this.lockOptions.method &&
+          this.lockOptions.method.toLowerCase() == 'fingerprint'
+        )
+          this.removeFingerprint();
         break;
       case 'fingerprint':
         this.lockByFingerprint();
@@ -83,13 +119,16 @@ export class LockPage {
   }
 
   private removeFingerprint(): void {
-    this.touchIdProvider.check().then(() => {
-      let lock = { method: 'disabled', value: null, bannedUntil: null };
-      this.configProvider.set({ lock });
-      this.checkLockOptions();
-    }).catch(() => {
-      this.checkLockOptions();
-    });
+    this.touchIdProvider
+      .check()
+      .then(() => {
+        let lock = { method: 'disabled', value: null, bannedUntil: null };
+        this.configProvider.set({ lock });
+        this.checkLockOptions();
+      })
+      .catch(() => {
+        this.checkLockOptions();
+      });
   }
 
   public lockByFingerprint(): void {
@@ -100,20 +139,27 @@ export class LockPage {
 
   private needsBackup() {
     let wallets = this.profileProvider.getWallets();
-    let singleLivenetWallet = wallets.length == 1 && wallets[0].network == 'livenet' && wallets[0].needsBackup;
-    let atLeastOneLivenetWallet = _.find(wallets, (w) => {
+    let singleLivenetWallet =
+      wallets.length == 1 &&
+      wallets[0].network == 'livenet' &&
+      wallets[0].needsBackup;
+    let atLeastOneLivenetWallet = _.find(wallets, w => {
       return w.network == 'livenet' && w.needsBackup;
     });
 
     if (singleLivenetWallet) {
-      this.needsBackupMsg = this.translate.instant('Backup your wallet before using this function');
+      this.needsBackupMsg = this.translate.instant(
+        'Backup your wallet before using this function'
+      );
       return true;
     } else if (atLeastOneLivenetWallet) {
-      this.needsBackupMsg = this.translate.instant('Backup all your wallets before using this function');
+      this.needsBackupMsg = this.translate.instant(
+        'Backup all your wallets before using this function'
+      );
       return true;
     } else {
       this.needsBackupMsg = null;
       return false;
     }
-  };
+  }
 }
